// src/graphql/type.interface.ts
export type Maybe<T> = T | null | undefined;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** A date-time string at UTC, such as 2019-12-03T09:54:33Z, compliant with the date-time format. */
  DateTime: any;
  /** The `JSONObject` scalar type represents JSON objects as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSONObject: any;
};

export enum ActiveCodeEnum {
  ACTIVATE = 'ACTIVATE',
  FORGOT = 'FORGOT',
}

export enum ActorTypeEnum {
  ADMIN = 'ADMIN',
  PARTNER = 'PARTNER',
  USER = 'USER',
}

export type AddressEntity = Node & {
  __typename?: 'AddressEntity';
  addressDetail?: Maybe<Scalars['String']>;
  addressName: Scalars['String'];
  contactName: Scalars['String'];
  contactPhone: Scalars['String'];
  createdAt: Scalars['DateTime'];
  deletedAt?: Maybe<Scalars['DateTime']>;
  id: Scalars['ID'];
  isDefault: Scalars['Boolean'];
  latitude: Scalars['Float'];
  longitude: Scalars['Float'];
  mapAddress: Scalars['String'];
  updatedAt: Scalars['DateTime'];
  userId: Scalars['String'];
};
// ...
export enum FileType {
  EXCEL = 'EXCEL',
  FILE = 'FILE',
  IMAGE = 'IMAGE',
  INVOICE = 'INVOICE',
  VIDEO = 'VIDEO',
}
// ...
export type Media = Node & {
  __typename?: 'Media';
  createdAt: Scalars['DateTime'];
  fileSize?: Maybe<Scalars['Int']>;
  fullOriginalUrl?: Maybe<Scalars['String']>;
  fullThumbUrl?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  isDeleted: Scalars['Boolean'];
  mimeType?: Maybe<Scalars['String']>;
  name: Scalars['String'];
  originalUrl: Scalars['String'];
  ownerId?: Maybe<Scalars['String']>;
  thumbUrl?: Maybe<Scalars['String']>;
  type: FileType;
  updatedAt: Scalars['DateTime'];
  videoUrl?: Maybe<Scalars['String']>;
};
// ...
export type Node = {
  id: Scalars['ID'];
};
// ...
export type UserEntity = Node & {
  __typename?: 'UserEntity';
  address?: Maybe<Scalars['String']>;
  avatar?: Maybe<Media>;
  avatarId?: Maybe<Scalars['String']>;
  birthday?: Maybe<Scalars['DateTime']>;
  certificate?: Maybe<Scalars['String']>;
  createdAt: Scalars['DateTime'];
  deletedAt?: Maybe<Scalars['DateTime']>;
  email?: Maybe<Scalars['String']>;
  fullname?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  isActive: Scalars['Boolean'];
  numberBooking: Scalars['Float'];
  numberMaintenance: Scalars['Float'];
  numberOrder: Scalars['Float'];
  phone: Scalars['String'];
  star: Scalars['Float'];
  totalBookings: Scalars['Float'];
  totalMaintenanceRequests: Scalars['Float'];
  totalOrders: Scalars['Float'];
  totalPayment: Scalars['Float'];
  totalSpending: Scalars['Float'];
  updatedAt: Scalars['DateTime'];
  userAddress?: Maybe<AddressEntity>;
};